+++ amazon/cognito/client/src/main/kotlin/org/http4k/connect/amazon/cognito/HttpCognito.kt
@@ -33,8 +33,9 @@ fun Cognito.Companion.Http(
 fun Cognito.Companion.Http(
     env: Map<String, String> = getenv(),
     http: HttpHandler = JavaHttpClient(),
-    clock: Clock = Clock.systemUTC()
-) = Http(Environment.from(env), http, clock)
+    clock: Clock = Clock.systemUTC(),
+    credentialsProvider: CredentialsProvider = CredentialsProvider.Environment(env)
+) = Http(Environment.from(env), http, clock, credentialsProvider)
 
 /**
  * Convenience function to create a Cognito from an http4k Environment
@@ -42,5 +43,6 @@ fun Cognito.Companion.Http(
 fun Cognito.Companion.Http(
     env: Environment,
     http: HttpHandler = JavaHttpClient(),
-    clock: Clock = Clock.systemUTC()
-) = Http(AWS_REGION(env), CredentialsProvider.Environment(env), http, clock)
+    clock: Clock = Clock.systemUTC(),
+    credentialsProvider: CredentialsProvider = CredentialsProvider.Environment(env)
+) = Http(AWS_REGION(env), credentialsProvider, http, clock)