+++ amazon/sts/client/src/main/kotlin/org/http4k/connect/amazon/sts/HttpSTS.kt
@@ -41,8 +41,10 @@ fun STS.Companion.Http(
 fun STS.Companion.Http(
     env: Map<String, String> = getenv(),
     http: HttpHandler = JavaHttpClient(),
-    clock: Clock = systemUTC()
-) = Http(Environment.from(env), http, clock)
+    clock: Clock = systemUTC(),
+    credentialsProvider: CredentialsProvider = CredentialsProvider.Environment(env)
+) = Http(Environment.from(env), http, clock, credentialsProvider)
+
 
 /**
  * Convenience function to create a STS from an http4k Environment
@@ -50,5 +52,6 @@ fun STS.Companion.Http(
 fun STS.Companion.Http(
     env: Environment,
     http: HttpHandler = JavaHttpClient(),
-    clock: Clock = systemUTC()
-) = Http(AWS_REGION(env), CredentialsProvider.Environment(env), http, clock)
+    clock: Clock = systemUTC(),
+    credentialsProvider: CredentialsProvider = CredentialsProvider.Environment(env)
+) = Http(AWS_REGION(env), credentialsProvider, http, clock)