+++ amazon/sts/client/src/test/kotlin/org/http4k/connect/amazon/sts/STSContract.kt
@@ -2,6 +2,7 @@ package org.http4k.connect.amazon.sts
 
 import org.http4k.connect.amazon.AwsContract
 import org.http4k.connect.amazon.core.model.ARN
+import org.http4k.connect.amazon.core.model.RoleSessionName
 import org.http4k.connect.successValue
 import org.http4k.core.HttpHandler
 import org.junit.jupiter.api.Assertions.assertTrue
@@ -22,7 +23,7 @@ abstract class STSContract(http: HttpHandler) : AwsContract() {
     fun `assume role`() {
         val result = sts.assumeRole(
             ARN.of("arn:aws:iam::169766454405:role/ROLETEST"),
-            UUID.randomUUID().toString(),
+            RoleSessionName.of(UUID.randomUUID().toString()),
             DurationSeconds = Duration.ofHours(1)
         )