+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v3.8.1.0
+- **http4k-connect-*** - Upgrade dependencies, including Kotlin to 1.5.30.
+- **http4k-connect-amazon-sts*** - Added convenience functions for Credential Providers.
+
 ### v3.8.0.0
+- **http4k-connect-*** - Upgrade dependencies.
 - **http4k-connect-amazon-sts*** - Support STS (refreshing) Credential providers including by WebIdentityToken.
 
 ### v3.7.0.0